move client factory into separate mixin class
Splits the factory interface away from the general
client management class so that it can be easily
removed or replaced as required.

Also removes the code for tracking scans in progress
and instead offers a callback to be executed when
add_disk() has finished. Driver implementations may
implement their own tracking with this. See ahci-driver
for an example.

Change-Id: I7886e42e8a27fe652319af0d56cb2ffb9e1e1102