Provide a pairing callback
This gives both static and dynamic clients a way to specify a callback
to be used when a client is paired with a device.

The Device_mgr class now supports an optional template parameter which
specifies the user-derived type from Partitioned_device so that
instances of this type can be created by the library.

Change-Id: I19b9a6f552f8f9d9d74eefb5751297c4498ede08