move pending queues from client to device
Per-client pending queues don't work when clients share a device
which can only handle a bounded number of requests in parallel.
Queues are now owned by a device and are requested on start-up
by the client. If a device does not provide a queue, requests
are dropped immediately when the device is busy.

Fixes issue #CD-20.

Change-Id: I82e147f0188dc90503ee71dd7b0548b655c3a1f8