use a factory class for the device manager
Instead of handing in the concrete classes that realise the
virtio interface and the partitioned device, provide a factory
class that contains functions to create these classes.

This gives us more flexibility how these classes are created.
For example, it is possible to use the standard implementation
and just set additional parameters before returning the instance.
It also allows to easily replace the device interface with one
that implements a completely different protocol.

Change-Id: I37616eb210008c7d82050c2ea572217783c97aae