make device type a template parameter of main classes
This changes the type of the device to a template parameter
for the manager and virtio client and consequently also for
the partition reader and inout memory helper.

When the base device type is a template parameter, then the
classes no longer directly depend on the Block_device::Device
type and devices may implement their own independent classes.

As a consequence it is now possible to implement to different
device types:

A simple device just may consist of one class and implement the
necessary function of a device as normal non-virtual functions.

A complex device that also wants to have support for partitions
continues to base its implementation on the Block_device::Device
interface and adds various implementations for it. The library
continues to provide a generic partition device for this use case.

More code restructuring will follow in subsequent changes to
to simplify handling of partitions.

Change-Id: Icf39e3b2c0b9562e4fee18a04912c28e70b2a300