make partition scanning optional
Partitions now will only be scanned, when the FACTORY class supplies
a create_partition() function with the correct parameters.

This is achieved using SFINAE. We provide two function templates
for scanning the disk. One uses return type deduction that depends
on the presence of the create_partition() function. The other can
always be resolved. An additional int vs long parameter ensures
that the compiler can correctly resolve the right function when
both templates are viable because the function is present.

Change-Id: I41df6bda1bd78d6ca35ca408b90108ab66cfa71f