make owner check responsibility of pending request
The item in the request queue explicitly saved the owner and
would check against it, when draining the queue. The queue
never calls the owner object but just forwards the owner again
to the callbacks. The disadvantage of this setup is that the
queue has to be aware of the type of the owner.

This change removes the owner member from the queue item and
keeps only the pending request. The pending request now needs
to keep track of the owner of the request and implement an
additional function that checks against a provided owner.

Change-Id: Ie2e96fd4654edcd68db94311b266a587dce88440