simplify handling of discard feature
With this commit the discard feature is no longer a mixin but
becomes a partial interface that can be added to the interface
definition of a device.

Partitioned_device and Virtio_client no longer need a special
mixin but automatically detect if the device has the discard
feature implemented and compile the appropriate code accordingly.

Partitioned_device requires that the device interface inherits
from Device_discard_feature in order to enable forwarding of
discard().

Virtio_client detects the presence of the discard feature via SFINAE.

Change-Id: I77213dd166d7632341e5e1623580a703e7d4baa1