Use local copy of negotiated features
The driver could modify the driver features in the shared configuration
space at any time which would influence the result of the newly computed
negotiated features. To avoid this problem, Virtio_client has a snapshot
of the negotiated features taken during the call to check_features(),
which the driver cannot silently modify. It is therefore preferable to
use the snapshot of the negotiated features.

Change-Id: I548262ba0500df0d915c98e15ad7ca5e84bb666a