Do not fail write-zeros req if unmap flag is set
For the VIRTIO_BLK_T_WRITE_ZEROES request the unmap flag might be set
by the driver. The device MAY deallocate the affected sectors but is
not obliged to do so. In fact, Linux will unconditionally set the unmap
flag for such sectors during ext4 file system initialization. Instead
we now look if the underlying driver supports the unmap operation. If
unsupported, the unmap flag will simply be ignored as the driver
visible result is always the same (all sectors are cleared to zero).

Change-Id: Ife01433b4fd01fa1750fb0c64f4ee449a1285f99