Break a self-referencing loop
The Partition_reader is a reference-counted object. After it is created
in scan_disk_partitions(), it is copy-captured into the callback for
read(). In read(), the callback is stored into the instance member
_callback. Since the instance references the callback and the callback
holds a reference to the instance, this represents a self-referencing
loop and a leak. Note that it can also prevent destruction of other
objects captured in the callback to scan_disk_partitions, such as
Connection and in turn Virtio_clients during Device_mgr destruction.

Fixes issue #CD-865.

Change-Id: If37952840b92f7992a3105bf0b056831396a151e
Requires-Change-Id: Ia477e1ab10f67bbca554e08432902caae4a114ad