Break self-referencing loop and defer destruction
The Partition_reader is a reference-counted object. After it is created
in scan_disk_partitions(), it is copy-captured into the callback for
read(). In read(), the callback is stored into the instance member
_callback. Since the instance references the callback and the callback
holds a reference to the instance, this represents a self-referencing
loop and a leak. Note that it can also prevent destruction of other
objects captured in the callback to scan_disk_partitions, such as
Connection and in turn Virtio_clients during Device_mgr destruction.

At the same time, it is necessary to defer the destruction of the
Partition_reader instance after the invocation of the callback
until the Partition_reader is not invoked. This is challenging as the
callback is typically invoked from the Partition_reader. This change
defers the destruction via scheduling an independent timeout and
capturing the last reference to the Partition_reader in its own
callback.

Fixes issue #CD-865.

Requires-Change-Id: Ia477e1ab10f67bbca554e08432902caae4a114ad
Requires-Change-Id: I969b5ea58f84a08ea8addb6377629f50c14958e0
Change-Id: I24b219c034341b1b7c5928a8a28a8f021d515685